Move VennDiagram into own file
